Fix handling of two edge loops
They are basically a vertex which has only one edge. The edge will be
removed.

I do not know if this may introduce new bugs, but it is necessary when
disabling merging of coplanar faces, as such edges were possibly merged
in that step. So far no new bugs have shown.