HEADER and ENDSEC added for the header section
When reading saved ifc files without any modifications we noticed that something went wrong. When we read the code we noticed that although the data section is surrounded by the DATA and ENDSEC tags, this didn't happen to the header. After consulting the standard for the header we saw that it also should, and that's the little modification proposed here. Right now, as the ReaderSTEP reads the header with a "\n" at the beginning and the end, when saving there's an extra blank line added. However, it is safer (it doesn't provoke any later error) as ReaderSTEP will maybe change this behaviour.